+++ History.md
@@ -1,6 +1,8 @@
 ## v.NEXT
 
-## v.1.2.1, ???
+* Split up `standard-minifiers` in separate CSS (`standard-minifiers-css`) and JS minifiers(`standard-minifiers-js`). `standard-minifiers` now acts as an umbrella package for these 2 minifiers.
+
+## v.1.2.1, 2015-Oct-26
 
 * `coll.insert()` now uses a faster (but cryptographically insecure)
   algorithm to generate document IDs when called outside of a method